Move the CheckFormModelMock class to external file for potential reuse
