+++ admin/spec/mocks/check-form-model-mock.js
@@ -1,0 +1,10 @@
+class CheckFormModelMock {
+  constructor (cb) {
+    this.done = cb
+  }
+  findOne () { return this }
+  lean () { return this }
+  exec () { return this.done() }
+}
+
+module.exports = CheckFormModelMock
